BIG CHANGES
This is a HUGE commit. It marks the starting of a new workflow. Here the pipeline already accepts a YAML samplesheet as input.

The pipelie has a new organization of files in directories and module names.

Used scripts and files in the pipeline will now to put under the "docker" directory for a proper versioning so each release has its own scripts version.

Stiil needs to finish documentation, implement possiblily of single and multiple samples, recheck the processes, outputs and reports and update docker images.