BIG Commit
New organization of modules. Pipeline going until JBROWSE. Now it is necessary to correctly add Resfinder annotation to the GFF, re-check the reports and make them the best up-to-date possible.

Check the possibiliy of drawing the genomic islands and prophages intersections and, of course, update the manual, documentation, output explanation and docker images.

Check the possibility of adding ICEfinder ... and, try to add prophage sequences to GFF? Check also the GenomicSqlite package.