Custom db and sql added
The use of custom databases and the creation of the SQLdb is not working for the single sample workflow.

It must be also implemented to the multi sample workflow and to be properly included into the main.nf and help message.

After that, I must create a R script and process to create a custom report for these custom annotations.

And finally, trigger the release, update docker images, update documentation with changes and explain the use of the pipeline and its results.